Relabel accounts
Discontent with the monikers used for classes, I renamed them (and  for consistency' sake I also renamed the documents that declare them).  It was a good opportunity to see Visual Studio's power for revisions.
This closes out Yellow Book 4.8 and, having made these somewhat sweeping changes, sets me up for work on 4.9.